fix: no more hyphens everywhere by upgrade Next 14 -> 15
There was a weird bug that only occured in production builds where some
tailwind class definitions for nextra were merged. The most visible side
effect is the md:hidden utility class was added to a list of many
classes that would set the ::after to content: '-'. This resulted in
hyphen characters appearing all over the UI.

This merging of classe only seemed to occur in the nextra theme
style.css chunk. This did not appear in dev builds `next dev`, only
`next build`

Here is an example of some of the corrupted styles:\
```css
  .x\:lg\:-translate-y-1\/2,
  .x\:lg\:after\:content-\[\"–\"\],
  .x\:lg\:border-b,
  .x\:lg\:right-full,
  .x\:lg\:top-1\/2,
  .x\:max-md\:break-all,
  .x\:max-md\:h-full,
  .x\:max-md\:hidden,
  .x\:max-md\:me-auto,
  .x\:max-md\:overflow-hidden,
  .x\:max-md\:sticky,
  .x\:max-sm\:hidden,
  .x\:md\:-mx-8,
  .x\:md\:hidden,
  .x\:md\:justify-start,
  .x\:md\:max-h-\[min\(calc\(100vh-5rem\)\,400px\)\]\!,
  .x\:md\:min-h-28,
  .x\:md\:px-12,
  .x\:md\:text-lg,
  .x\:md\:text-sm,
  .x\:md\:w-64,
  .x\:md\:w-\[576px\] {
    @media (width >= 64rem) {
      &:after {
        /*! --tw-content:"–"; */content:var(--tw-content)
      }
    }
  }
  .x\:2xl\:-mx-24,
  .x\:ltr\:rotate-90,
  .x\:xl\:me-\[calc\(50\%-50vw\)\],
  .x\:xl\:ms-\[calc\(50\%-50vw\+16rem\)\] {
    &:where(:dir(ltr),[dir=ltr],[dir=ltr] *) {
      rotate:90deg
    }
  }
```